Extract mint item method
